// src/api/models/GeneralStatOutput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WinLoss } from './WinLoss';
import {
    WinLossFromJSON,
    WinLossFromJSONTyped,
    WinLossToJSON,
    WinLossToJSONTyped,
} from './WinLoss';
import type { General } from './General';
import {
    GeneralFromJSON,
    GeneralFromJSONTyped,
    GeneralToJSON,
    GeneralToJSONTyped,
} from './General';
import type { GeneralStatPlayerWL } from './GeneralStatPlayerWL';
import {
    GeneralStatPlayerWLFromJSON,
    GeneralStatPlayerWLFromJSONTyped,
    GeneralStatPlayerWLToJSON,
    GeneralStatPlayerWLToJSONTyped,
} from './GeneralStatPlayerWL';

/**
 * 
 * @export
 * @interface GeneralStatOutput
 */
export interface GeneralStatOutput {
    /**
     * 
     * @type {General}
     * @memberof GeneralStatOutput
     */
    general: General;
    /**
     * 
     * @type {Array<GeneralStatPlayerWL>}
     * @memberof GeneralStatOutput
     */
    stats: Array<GeneralStatPlayerWL>;
    /**
     * 
     * @type {WinLoss}
     * @memberof GeneralStatOutput
     */
    total: WinLoss;
}



/**
 * Check if a given object implements the GeneralStatOutput interface.
 */
export function instanceOfGeneralStatOutput(value: object): value is GeneralStatOutput {
    if (!('general' in value) || value['general'] === undefined) return false;
    if (!('stats' in value) || value['stats'] === undefined) return false;
    if (!('total' in value) || value['total'] === undefined) return false;
    return true;
}

export function GeneralStatOutputFromJSON(json: any): GeneralStatOutput {
    return GeneralStatOutputFromJSONTyped(json, false);
}

export function GeneralStatOutputFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeneralStatOutput {
    if (json == null) {
        return json;
    }
    return {
        
        'general': GeneralFromJSON(json['general']),
        'stats': ((json['stats'] as Array<any>).map(GeneralStatPlayerWLFromJSON)),
        'total': WinLossFromJSON(json['total']),
    };
}

export function GeneralStatOutputToJSON(json: any): GeneralStatOutput {
    return GeneralStatOutputToJSONTyped(json, false);
}

export function GeneralStatOutputToJSONTyped(value?: GeneralStatOutput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'general': GeneralToJSON(value['general']),
        'stats': ((value['stats'] as Array<any>).map(GeneralStatPlayerWLToJSON)),
        'total': WinLossToJSON(value['total']),
    };
}